pyproject.toml
[project]
name = "rag-project"
version = "1.0.0"
description = "Rag Weg"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi[standard]>=0.128.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.3",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["app"]

[tool.setuptools]
package-dir = { "" = "app" }
